bug fix: userid should be long not int(int32).

+++ Assets/NormalScripts/Network/NetworkMsg.cs
@@ -27,7 +27,7 @@ public class NetworkMsg : object
 
 public class LocalAuthMsg: NetworkMsg
 {
-    public int UserID;
+    public long UserID;
     public string UserName;
     public string Password;
 }
